v4.0.0: Setup workflow with gulp
'gulp build' now launches browserify plus babelify to build our JS
bundle. For some reason babelify was throwing an error on ES6 imports
with an npm task.
Added a React component that renders a stub text when we request data
with the button
Added step npm install on the Readme